refactor(core): extract analysis request owner
